requirements.txt
playwright==1.43.0
pandas==2.2.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
requests==2.31.0
Jinja2==3.1.3
